Polycube snakes and strips: further developments
My lovely wife Sam points out yet another species of snake or ouroboros
in three dimensions: Besides enclosing a full-on hole or "cavity,"
(as in A357083 and A355966) we could imagine a snake with the topology
of a donut, such that an infinitely long string passing through the "hole"
cannot be wiggled to the outside except by passing through a corner
of the snake. For ouroboroi, we could even distinguish those which
require the string to pass through a cube itself (ones where the
ouroboros encircles a hole) from the others (ones where the ouroboros
doubles back on itself and the doubled path itself encloses a hole
that lies "outside" the ouroboros' coil).

One of the chiral polycube ouroboroi with n=10 has donut topology;
the other does not.

There is a quite beautifully threefold-symmetric 12-cube ouroboros
which encloses a cavity.