Update Godbolt links in "boom goes the coroutines"
This old post was a little cavalier about string literals; a literal
that is bound to a `std::string` in the first snippet quietly turns into
a static-storage string literal in the `string_view` snippet. Add a
UDL suffix `s` to prevent that from happening, and give a Godbolt link
so skeptics can just _see_ what I meant. Hat tip to Hui Xie.