Oops, Microsoft STL's `string` has no pointer-to-self
Thanks to HN commenter "cmovq" for raising the alarm. Microsoft's
`string`, like libc++'s, is trivially relocatable and contains no
pointer-to-self. When I wrote the now-deleted sentences, I somehow
thought it did. I even read my own table wrong! There's a checkmark
in the right place in the table in
https://quuxplusone.github.io/blog/2019/02/20/p1144-what-types-are-relocatable/
I just managed to misread it.